+++ ewsgui/internal/functions/Method8.ps1
@@ -6,6 +6,15 @@
     .DESCRIPTION
     Method to create a custom folder in mailbox's Root.
     
+    .PARAMETER ClientID
+    String parameter with the ClientID (or AppId) of your AzureAD Registered App.
+
+    .PARAMETER TenantID
+    String parameter with the TenantID your AzureAD tenant.
+
+    .PARAMETER ClientSecret
+    String parameter with the Client Secret which is configured in the AzureAD App.
+    
     .EXAMPLE
     PS C:\> Method8
     Method to create a custom folder in mailbox's Root.
@@ -13,11 +22,18 @@
     #>
     [CmdletBinding()]
     param(
-        # Parameters
+        [String] $ClientID,
+
+        [String] $TenantID,
+
+        [String] $ClientSecret
     )
+    
     if ( $txtBoxFolderID.Text -ne "" )
     {
         $statusBarLabel.text = "Running..."
+        Test-StopWatch -Service $service -ClientID $ClientID -TenantID $TenantID -ClientSecret $ClientSecret
+
         $folder = new-object Microsoft.Exchange.WebServices.Data.Folder($service)
         $folder.DisplayName = $txtBoxFolderID.Text
         $folder.Save([Microsoft.Exchange.WebServices.Data.WellKnownFolderName]::MsgFolderRoot)